+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2023-10-05
+
 ### Added
 
 - `addonLeft` and `addonRight` props now also accept a function to render a custom component
@@ -323,7 +325,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.0.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.1.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -348,6 +350,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[2.1.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.6.0...2.0.0
 [1.6.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.4.0...1.5.0